requirements.txt
networkx
ipykernel 
kubernetes
prometheus_api_client
numpy
graphviz
langgraph
langchain
jupyter
openai
python-dotenv
langgraph-checkpoint-sqlite
langchain_community
langchain_core
langchain_openai>=0.3.9rc1
psycopg2
litemcp
